pyproject.toml
[tool.poetry]
name = "quantum-edge"
version = "0.1.0"
description = "Autonomous AI-powered day trading platform with 7-agent architecture"
authors = ["Ben Camara"]
readme = "README.md"
packages = [
    { include = "quantum_edge" },
    { include = "agents" },
    { include = "coordinator" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pydantic = "^2.10"
pydantic-settings = "^2.7"
redis = {version = "^5.2", extras = ["hiredis"]}
asyncpg = "^0.30"
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
alpaca-py = "^0.33"
httpx = "^0.28"
fastapi = "^0.115"
uvicorn = {version = "^0.34", extras = ["standard"]}
structlog = "^24.4"
prometheus-client = "^0.21"
transformers = "^4.47"
torch = "^2.5"
scikit-learn = "^1.6"
hmmlearn = "^0.3"
arch = "^7.2"
pandas = "^2.2"
numpy = "^2.1"
pandas-ta = ">=0.4.0b0"
feedparser = "^6.0"
beautifulsoup4 = "^4.12"
aiohttp = "^3.11"
orjson = "^3.10"
tenacity = "^9.0"
croniter = "^5.0"
streamlit = "^1.41"
python-dotenv = "^1.0"
PyJWT = "^2.9"
bcrypt = "^4.2"
python-multipart = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-cov = "^6.0"
ruff = "^0.8"
mypy = "^1.13"
fakeredis = "^2.26"
factory-boy = "^3.3"
testcontainers = {version = "^4.9", extras = ["redis", "postgres"]}
freezegun = "^1.4"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring Docker services",
    "slow: Slow tests (backtesting, model training)",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "A", "COM", "C4", "DTZ", "ISC", "PIE", "PT", "RSE", "RET", "SIM", "TID", "TCH", "ARG", "PTH", "ERA"]
ignore = ["ANN101", "ANN102", "ANN401", "S101", "COM812", "ISC001"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
